Move ReturnToSpController to Redirect module
**Why**: Since it's now an extending controller of the base RedirectController
